third_party/WebKit/LayoutTests/external/wpt/css/css-animations/CSSAnimation-animationName.tentative-expected.txt

This is a testharness.js-based test.
FAIL Animation name makes keyframe rule assert_equals: Animation name matches keyframes rule name expected (string) "xyz" but got (undefined) undefined
FAIL Escaped animation name assert_equals: Escaped animation name matches keyframes rule name expected (string) "xyz" but got (undefined) undefined
FAIL Animation name with hex-escape assert_equals: Hex-escaped animation name matches keyframes rule name expected (string) "xyz" but got (undefined) undefined
Harness: the test ran to completion.
